[Http] Refactor UriFactory.
