+++ tests/Unit/Http/Message/Uri/Factory/UriFactoryTest.php
@@ -13,9 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Tests\Unit\Http\Message\Uri\Factory;
 
-use Valkyrja\Http\Message\Header\Header;
 use Valkyrja\Http\Message\Throwable\Exception\InvalidArgumentException;
-use Valkyrja\Http\Message\Uri\Data\HostPortAccumulator;
 use Valkyrja\Http\Message\Uri\Enum\Scheme;
 use Valkyrja\Http\Message\Uri\Factory\UriFactory;
 use Valkyrja\Http\Message\Uri\Psr\Uri as PsrUri;
@@ -82,166 +80,6 @@ class UriFactoryTest extends TestCase
         UriFactory::fromString('//');
     }
 
-    public function testMarshalUriFromServer(): void
-    {
-        $host     = 'https://www.host.com';
-        $port     = 70;
-        $path     = '/path';
-        $query    = 'query=test';
-        $fragment = 'fragment';
-        $server   = [
-            'HTTPS'        => 'on',
-            'SERVER_NAME'  => $host,
-            'SERVER_PORT'  => (string) $port,
-            'REQUEST_URI'  => "$path#$fragment",
-            'QUERY_STRING' => $query,
-        ];
-        $headers  = [];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        self::assertSame(Scheme::HTTPS, $uri->getScheme());
-        self::assertSame($host, $uri->getHost());
-        self::assertSame($port, $uri->getPort());
-        self::assertSame($path, $uri->getPath());
-        self::assertSame($query, $uri->getQuery());
-        self::assertSame($fragment, $uri->getFragment());
-    }
-
-    public function testMarshalEmptyServerAndHeaders(): void
-    {
-        $blankAccumulator = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders($blankAccumulator, [], []);
-
-        self::assertEmpty($blankAccumulator->host);
-        self::assertNull($blankAccumulator->port ?? null);
-    }
-
-    public function testMarshalHostAndPortFromHeaders(): void
-    {
-        $host = 'https://www.host.com';
-        $port = 70;
-
-        $hostHeaderAccumulator = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders($hostHeaderAccumulator, [], ['host' => new Header('Host', "$host:$port")]);
-
-        self::assertSame($host, $hostHeaderAccumulator->host);
-        self::assertSame($port, $hostHeaderAccumulator->port);
-    }
-
-    public function testServerHostAndPortAccumulator(): void
-    {
-        $host = 'https://www.host.com';
-        $port = 70;
-
-        $serverAccumulator = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders(
-            $serverAccumulator,
-            [
-                'SERVER_NAME' => $host,
-                'SERVER_PORT' => (string) $port,
-            ],
-            []
-        );
-
-        self::assertSame($host, $serverAccumulator->host);
-        self::assertSame($port, $serverAccumulator->port);
-    }
-
-    public function testIp6FallbackAccumulator(): void
-    {
-        $port = 70;
-
-        $ip6FallbackAccumulator = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders(
-            $ip6FallbackAccumulator,
-            [
-                'SERVER_NAME' => '[FE80::0202:B3FF:FE1E:8329]',
-                'SERVER_PORT' => (string) $port,
-                'SERVER_ADDR' => 'FE80::0202:B3FF:FE1E:8329',
-            ],
-            []
-        );
-
-        self::assertSame('[FE80::0202:B3FF:FE1E:8329]', $ip6FallbackAccumulator->host);
-        self::assertSame($port, $ip6FallbackAccumulator->port);
-    }
-
-    public function testIp6FallbackAccumulatorNullPort(): void
-    {
-        $ip6FallbackAccumulatorNullPort = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders(
-            $ip6FallbackAccumulatorNullPort,
-            [
-                'SERVER_NAME' => '[FE80::0202:B3FF:FE1E:8329]',
-                'SERVER_PORT' => '90',
-                'SERVER_ADDR' => 'FE80::0202:B3FF:FE1E:8329:90',
-            ],
-            []
-        );
-
-        self::assertSame('[FE80::0202:B3FF:FE1E:8329:90]', $ip6FallbackAccumulatorNullPort->host);
-        self::assertNull($ip6FallbackAccumulatorNullPort->port);
-    }
-
-    public function testMarshalRequestUri(): void
-    {
-        $unencodedUrl = UriFactory::marshalRequestUri([
-            'IIS_WasUrlRewritten' => '1',
-            'UNENCODED_URL'       => $unencodedUrlExpected = '/unencodedUrl',
-        ]);
-
-        self::assertSame($unencodedUrlExpected, $unencodedUrl);
-
-        $httpXRewriteUrl = UriFactory::marshalRequestUri([
-            'HTTP_X_REWRITE_URL' => $httpXRewriteUrlExpected = '/httpXRewriteUrl',
-        ]);
-
-        self::assertSame($httpXRewriteUrlExpected, $httpXRewriteUrl);
-
-        $httpXOriginalUrl = UriFactory::marshalRequestUri([
-            'HTTP_X_ORIGINAL_URL' => $httpXOriginalUrlExpected = '/httpXOriginalUrl',
-        ]);
-
-        self::assertSame($httpXOriginalUrlExpected, $httpXOriginalUrl);
-
-        $requestUri = UriFactory::marshalRequestUri([
-            'REQUEST_URI' => $requestUriExpected = '/requestUri',
-        ]);
-
-        self::assertSame($requestUriExpected, $requestUri);
-
-        $origPathInfo = UriFactory::marshalRequestUri([
-            'ORIG_PATH_INFO' => $origPathInfoExpected = '/origPathInfo',
-        ]);
-
-        self::assertSame($origPathInfoExpected, $origPathInfo);
-
-        $origPathInfoEmpty = UriFactory::marshalRequestUri([
-            'ORIG_PATH_INFO' => '',
-        ]);
-
-        self::assertSame('/', $origPathInfoEmpty);
-
-        $noMatchedServerParams = UriFactory::marshalRequestUri([]);
-
-        self::assertSame('/', $noMatchedServerParams);
-    }
-
-    public function testStripQueryString(): void
-    {
-        $pathString      = UriFactory::stripQueryString($path = '/path');
-        $pathStringWithQ = UriFactory::stripQueryString('/path?query=string');
-
-        self::assertSame($path, $pathString);
-        self::assertSame($path, $pathStringWithQ);
-    }
-
     public function testFromPsrArray(): void
     {
         $stream = new Uri(
@@ -267,25 +105,6 @@ class UriFactoryTest extends TestCase
         self::assertSame($path, $fromPsr->getPath());
         self::assertSame($query, $fromPsr->getQuery());
         self::assertSame($fragment, $fromPsr->getFragment());
-    }
-
-    public function testGetHeader(): void
-    {
-        $headers = [
-            'array'  => new Header('Array', 'test1, test2'),
-            'string' => new Header('String', 'string'),
-        ];
-
-        self::assertSame('string', UriFactory::getHeader('String', $headers));
-        self::assertSame('test1, test2', UriFactory::getHeader('Array', $headers));
-    }
-
-    public function testGetHeaderWithDefault(): void
-    {
-        $headers = [];
-
-        self::assertSame('', UriFactory::getHeader('NonExistent', $headers));
-        self::assertSame('default', UriFactory::getHeader('NonExistent', $headers, 'default'));
     }
 
     public function testFilterScheme(): void
@@ -501,60 +320,6 @@ class UriFactoryTest extends TestCase
         self::assertSame('', $fromPsr->getPassword());
     }
 
-    public function testMarshalUriFromServerWithXForwardedProto(): void
-    {
-        $server = [
-            'SERVER_NAME' => 'example.com',
-            'SERVER_PORT' => '80',
-            'REQUEST_URI' => '/path',
-        ];
-        $headers = [
-            'x-forwarded-proto' => new Header('X-Forwarded-Proto', 'https'),
-        ];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        self::assertSame(Scheme::HTTPS, $uri->getScheme());
-    }
-
-    public function testMarshalUriFromServerWithHttpsOff(): void
-    {
-        $server = [
-            'HTTPS'       => 'off',
-            'SERVER_NAME' => 'example.com',
-            'SERVER_PORT' => '80',
-            'REQUEST_URI' => '/path',
-        ];
-        $headers = [];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        self::assertSame(Scheme::HTTP, $uri->getScheme());
-    }
-
-    public function testMarshalUriFromServerWithQueryStringLeadingQuestionMark(): void
-    {
-        $server = [
-            'SERVER_NAME'  => 'example.com',
-            'REQUEST_URI'  => '/path',
-            'QUERY_STRING' => '?key=value',
-        ];
-        $headers = [];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        self::assertSame('key=value', $uri->getQuery());
-    }
-
-    public function testMarshalRequestUriWithFullUrl(): void
-    {
-        $result = UriFactory::marshalRequestUri([
-            'REQUEST_URI' => 'http://example.com/path',
-        ]);
-
-        self::assertSame('/path', $result);
-    }
-
     public function testValidateFragment(): void
     {
         // validateFragment is currently empty but should not throw for any input
@@ -587,49 +352,6 @@ class UriFactoryTest extends TestCase
         self::assertTrue(true); // If we reach here, no exception was thrown
     }
 
-    public function testMarshalUriFromServerWithEmptyHost(): void
-    {
-        $server  = [
-            'REQUEST_URI' => '/path',
-        ];
-        $headers = [];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        // With no host information, host should be empty
-        self::assertSame('', $uri->getHost());
-        self::assertNull($uri->getPort());
-        self::assertSame('/path', $uri->getPath());
-    }
-
-    public function testMarshalHostAndPortFromHeadersWithHostNoPort(): void
-    {
-        $accumulator = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders(
-            $accumulator,
-            [],
-            ['host' => new Header('Host', 'example.com')]
-        );
-
-        self::assertSame('example.com', $accumulator->host);
-        self::assertNull($accumulator->port);
-    }
-
-    public function testMarshalHostAndPortFromHeadersWithServerNameNoPort(): void
-    {
-        $accumulator = new HostPortAccumulator();
-
-        UriFactory::marshalHostAndPortFromHeaders(
-            $accumulator,
-            ['SERVER_NAME' => 'example.com'],
-            []
-        );
-
-        self::assertSame('example.com', $accumulator->host);
-        self::assertNull($accumulator->port);
-    }
-
     public function testFilterPathWithValidPath(): void
     {
         // filterPath calls validatePath internally, then normalizes slashes
@@ -650,32 +372,6 @@ class UriFactoryTest extends TestCase
         $this->expectException(InvalidQueryException::class);
 
         UriFactory::filterQuery('query=value#fragment');
-    }
-
-    public function testMarshalUriFromServerWithNoQueryString(): void
-    {
-        $server = [
-            'SERVER_NAME' => 'example.com',
-            'REQUEST_URI' => '/path',
-        ];
-        $headers = [];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        self::assertSame('', $uri->getQuery());
-    }
-
-    public function testMarshalUriFromServerWithNoFragment(): void
-    {
-        $server = [
-            'SERVER_NAME' => 'example.com',
-            'REQUEST_URI' => '/path',
-        ];
-        $headers = [];
-
-        $uri = UriFactory::marshalUriFromServer($server, $headers);
-
-        self::assertSame('', $uri->getFragment());
     }
 
     public function testToStringWithOnlyPath(): void
